requirements.txt
# Mercedes W222 OBD Scanner Requirements - Production Ready

# GUI Framework
customtkinter>=5.2.0
pillow>=10.0.0

# OBD Communication
pyserial>=3.5
obd>=0.7.1

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0
configparser>=5.3.0

# Logging and Utils
loguru>=0.7.0
python-dateutil>=2.8.0

# Database (SQLite is built-in)
# Additional database utilities
sqlalchemy>=2.0.0

# Web Framework for API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
python-multipart>=0.0.6

# AI and Machine Learning
anthropic>=0.7.0
openai>=1.3.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Export and Reporting
reportlab>=3.6.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Security and Licensing
cryptography>=41.0.0
requests>=2.31.0
psutil>=5.9.0
pyjwt>=2.8.0

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Production Deployment
gunicorn>=21.2.0
redis>=5.0.0
celery>=5.3.0

# Monitoring and Observability
prometheus-client>=0.19.0
structlog>=23.2.0

# Additional Utilities
aiofiles>=23.2.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
